pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.118.3",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "openai>=2.3.0",
    "pymongo>=4.15.3",
    "pypdf>=6.1.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
